refactor(agent): separate ToolConfig and improve ToolState interfaces
* Extract ToolConfig into standalone files in both Python and TS
* Convert ToolState into interface and add implementation class in TS
